services/lightweight-classifier/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
scikit-learn==1.3.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
pydantic==2.5.0
joblib==1.3.2
